Ejercicios extra

+++ src/ej09.py
@@ -1,2 +1,2 @@
 # Hacemos un input de los tres numeros y los metemos dentro del print para evitar trabajar con variables
-print('El total de esos tres números es: ' + str(int(input('Introduzca el primer número: ')) + int(input('Introduzca el segundo número: ')) + int(input('Introduzca el tercer número: '))))
+print('El total de esos tres números es: ' + str(float(input('Introduzca el primer número: ')) + float(input('Introduzca el segundo número: ')) + float(input('Introduzca el tercer número: '))))
